Adds asymmetric quadratic parametric model
Introduces a new parametric model, AsymQuadModel, for fitting asymmetric quadratic functions.
Extends the base ParametricModel class to include functionality for handling extrapolation,
including clipping input 'x' values to the fitting range when extrapolation is disabled.
Also adds functionality to store the bounds, and to serialize and deserialize the class to
and from dictionaries.